% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{plot.scm_spillover}
\alias{plot.scm_spillover}
\title{Plot method for scm_spillover objects}
\usage{
\method{plot}{scm_spillover}(x, type = "effects", ...)
}
\arguments{
\item{x}{An object of class scm_spillover}

\item{type}{Type of plot: "effects" or "all"}

\item{...}{Additional arguments passed to plotting functions}
}
\value{
A ggplot object (invisibly for type="all")
}
\description{
Plot method for scm_spillover objects
}
