% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{print.scm_spillover}
\alias{print.scm_spillover}
\title{Print method for scm_spillover objects}
\usage{
\method{print}{scm_spillover}(x, ...)
}
\arguments{
\item{x}{An object of class scm_spillover}

\item{...}{Additional arguments (not used)}
}
\value{
Invisibly returns the input object
}
\description{
Print method for scm_spillover objects
}
