% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scm_core.R
\name{scm_batch}
\alias{scm_batch}
\title{Batch SCM for all units}
\usage{
scm_batch(Y, verbose = FALSE)
}
\arguments{
\item{Y}{N x T matrix of outcomes}

\item{verbose}{Print progress messages}
}
\value{
List with intercepts (a) and weight matrix (B)
}
\description{
Computes synthetic control weights for all units
}
