% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spillover.R
\name{sp_andrews_spillover}
\alias{sp_andrews_spillover}
\title{Test for spillover effects}
\usage{
sp_andrews_spillover(Y0, Y1, A, alpha_sig = 0.05)
}
\arguments{
\item{Y0}{N x T pre-treatment matrix}

\item{Y1}{N x 1 post-treatment vector}

\item{A}{N x k spillover structure matrix}

\item{alpha_sig}{Significance level}
}
\value{
List with test results
}
\description{
Tests whether spillover effects are significant
}
