% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spillover.R
\name{sp_estimation}
\alias{sp_estimation}
\title{Spillover-adjusted SCM estimation}
\usage{
sp_estimation(Y_pre, Y_post, A, verbose = FALSE)
}
\arguments{
\item{Y_pre}{N x T pre-treatment matrix}

\item{Y_post}{N x S post-treatment matrix}

\item{A}{N x k spillover structure matrix}

\item{verbose}{Print progress messages}
}
\value{
List containing treatment effects and other results
}
\description{
Estimates treatment effects accounting for spillovers
}
