% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{summary.scm_spillover}
\alias{summary.scm_spillover}
\title{Summary method for scm_spillover objects}
\usage{
\method{summary}{scm_spillover}(object, ...)
}
\arguments{
\item{object}{An object of class scm_spillover}

\item{...}{Additional arguments (not used)}
}
\value{
Invisibly returns a data frame with detailed results
}
\description{
Summary method for scm_spillover objects
}
