% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scm_core.R
\name{vanilla_scm}
\alias{vanilla_scm}
\title{Standard SCM without spillovers}
\usage{
vanilla_scm(Y_pre, Y_post)
}
\arguments{
\item{Y_pre}{N x T pre-treatment matrix}

\item{Y_post}{N x S post-treatment matrix}
}
\value{
Synthetic control values for all periods
}
\description{
Computes standard synthetic control for comparison
}
